feat: update code

+++ models/items.go
@@ -3,5 +3,5 @@ package models
 
 type Item interface {
 	GetName() string
-	GetHarga() int64
+	GetHarga() int
 }
