+++ models/makanan.go
@@ -2,14 +2,14 @@ package models
 
 type Makanan struct {
 	Name string
-	Harga int64
+	Harga int
 }
 
 
 func (m Makanan) GetName()string {return m.Name}
-func (m Makanan) GetHarga()int64 {return m.Harga}
+func (m Makanan) GetHarga()int {return m.Harga}
 
-func NewMakanan(name string, harga int64)Makanan{
+func NewMakanan(name string, harga int)Makanan{
 	return Makanan{Name: name, Harga: harga}
 }
